requirements.txt
fastapi==0.115.5
fastapi-cli==0.0.5
keras==2.11.0
Keras-Preprocessing==1.1.2
numpy==2.0.2
requests==2.32.3
requests-oauthlib==1.3.1
tensorflow-estimator==2.11.0
tensorflow==2.11.0
tf-explain==0.3.1
urllib3==1.26.19
pillow==10.4.0